More Work on Expanding AI

+++ playerPlacement.m
@@ -1,10 +1,10 @@
 function [battleground,playerX,playerY] = playerPlacement(playerNumber,Y,X,battleground,playerHealth)
-    for en = uint8(1:playerNumber)
-        playerY = int16(randi(Y));
-        playerX = int16(randi(X)); 
+    for en = (1:playerNumber)
+        playerY = (randi(Y));
+        playerX = (randi(X)); 
         if battleground(playerY,playerX) ~= 0
-            playerY = int16(randi(Y));
-            playerX = int16(randi(X));
+            playerY = (randi(Y));
+            playerX = (randi(X));
         end
         battleground(playerY,playerX) = playerHealth;
     end
